tast-test: Stabilize RecentApps test
On the CQ, the informational test RecentApps is being flaky on various
boards due to different reasons.

At the beginning of the test, the test uses the common method
SetUpLauncherTest to open the launcher and try to stabilizing the number
of apps available. However, if the installed apps are not done
installing, we may not be able to show recent apps on the first launcher
open. The RecentApps test expects system web apps to be installed by the
time the launcher is opened, so explicitly wait for their installation
before calling SetUpLauncherTest (and opening the launcher).

Another common failure is that the recent apps order is not correct.
The flow of the test is at glance:
1  Open launcher
2  Install an app and verify it is first on RecentApps.
3  Open another app and verify RecentApps elements reorder.
4  Open the recently installed app and verify reorder again.
5  Uninstall recent app and verify it does not exists on RecentApps
   anymore.

The failure is usually on step #3 where the recent apps does not appear
to have reordered sometimes. As a speculative fix, wait for the app
window to become active before moving on to verify the order.

BUG=chromium:1364766
TEST=tast run DUT launcher.RecentApps*

Change-Id: I1b63e385eb99c93bf97278902be3889d3f18b95c