tast-test: Dismiss sort nudge to show recent apps
When opening the launcher, the usual flow shows partially the continue
section of which Recent Apps is part of. A privacy nudge shows where the
continue section is usually in place. After the user acknowledge the
notice, the UI for the continue section and recent apps is shown.

However, sometimes if the zero state search results have a delay at
retrieval, the sort nudge is shown instead. The Sort nudge will not
dismiss unless the user explicitly acknowledge it, and in turn, will
block the continue section to show with recent apps.

For this case, add some logic to the tast test to dismiss the sort nudge
if it was shown before the RecentApps could be fetched.

BUG=chromium:1364766
TEST=tast run DUT launcher.RecentApps*

Change-Id: I1b63e385eb99c93bf97278902be3889d3f18b95c