tast-tests: Parameterize test to use VK
This CL also changes the test slightly. Instead of logging in with
pin, it now logs in with password because for VK there is a test that
you cannot delete the credential you logged in with. This will be gone
once we have USS only backing store.

BUG=b:240594146
TEST=CQ Dry Passes

Change-Id: I4cfa7493bf252c93550197e9da4e88ced7531aa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3920272
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>