Wait for window in launcher.RemoveSuggestedSearchResult
The test opens a browser window (that searches google using a test
query), and closes it immediately. The test was flaky when getting
the browser window using GetActiveWindow (after waiting for a node
associated with the browser window). It looks like the browser
window opens, but is not considered active due to a location
permission prompt. Note that GetActiveWindow() is used to get the
browser window so the test can close it - this is not actually
necessary for the test to work (and is not an important part of
the test), so skip the step all together.

BUG=chromium:1368375
TEST=tast run <dut> launcher.RemoveSuggestedSearchResult*

Change-Id: I01346f75be16d6585dafb8b7d9a6fff3c2111041