Wait for window in launcher.RemoveSuggestedSearchResults
The test opens a browser window (that searches google using a test
query), and closes it immediately. The test was flaky when getting
the browser window using GetActiveWindow (after waiting for a node
associated with the browser window). It looks like the browser
window opens, but is not considered active due to a location
permission prompt. Instead, use ash.WaitForAnyWindowWithTitle to
get the browser window opened by the test. It simplifies the code
that waits for the browser window to show, and it does not require
the browser window to be active (which is not required for the
test).

BUG=chromium:1368375
TEST=tast run <dut> launcher.RemoveSuggestedSearchResults*

Change-Id: I01346f75be16d6585dafb8b7d9a6fff3c2111041