tast-tests: Close Files app window to enable single window on unmount
During unmounting of the SMB file shares in filemanager.SMBPassword the
test starts a foreground window and calls the fileManagerPrivate API to
unmount the SMB file share. If the previous window isn't closed, the
launch doesn't re-use the old window and it can't find a unique target
to connect to (there are 2 foreground windows). This closes the first
window to ensure uniqueness on unmount.

BUG=chromium:1360263
TEST=tast run filemanager.SMBPassword*

Change-Id: I7139d985b23b1cf97d20f7674e2bd0f3c418f4f9