typec: Explicitly close SSH link on disconnect
The host occasionally has trouble re-establishing an SSH link with the
DUT on CC-on reconnection (despite the DUT and servod both being
active). This could be attributed to the DUT not having it's SSH link
gracefully torn down (since the original disconnection was performed by
switching the CC line off on the servo).

To avoid this, explicitly close the host's SSH connection to the DUT
when we perform a CC-off disconnect.

BUG=b:249331815
TEST=typec.Basic & typec.HpdWake pass 30/30 times on a crosfleet volteer
device.

Change-Id: Icb988ba174b8b469fbb3b1b49a012db53b3ca0c3