ui: Add mouse drag actions to multiple CUJ tests
Update DesksCUJ, GoogleSheetsCUJ, GoogleSlidesCUJ, and QuickCheckCUJ to
include mouse dragging actions to ensure mouse drag metrics are
collected. Each test now clicks in the center of the screen, and drags
the mouse out to the left and right side, thereby highlighting elements
on the page. The mouse moves back to the center of the screen, limiting
the side effects of this drag movement.

BUG=b:241841511
TEST=tast run <DUT> ui.DesksCUJ ui.GoogleSheetsCUJ ui.GoogleSlidesCUJ
ui.QuickCheckCUJ, and ensure that
EventLatency.MouseDragged.TotalLatency is reported in
results-chart.json.

Change-Id: Ic37b3ba27b6a53909d1b3bd5cb2dc67b310cf42f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3923906
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>