tast-tests: Fix policy.ScreenCaptureAllowed
The tab order of the media picker changed and now the "Entire Screen"
tab is not the first anymore. The ScreenCaptureAllowed tast test was
assuming the "Entire Screen" tab is the first. With this change, the
test now specifically looks for the desired tab and clicks on it,
instead of assuming it's the default one.

BUG=b:245757370
TEST=tast run <device> policy.ScreenCaptureAllowed*

Change-Id: I2698d720a2f423ad1003b3ed266b6296a8e559e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3924340
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Tested-by: Maria Petrisor <mpetrisor@chromium.org>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>