tast-tests: Fix tast.policy.ScreenCaptureAllowed
The tab order of the media picker changed and now the "Entire Screen"
tab is not the first anymore. The ScreenCaptureAllowed tast test was
assuming the "Entire Screen" tab is the first. With this change, the
test now specifically looks for the desired tab and clicks on it,
instead of assuming it's the default one.

BUG=b:245757370
TEST=tast run <device> policy.policy.ScreenCaptureAllowed*

Change-Id: I2698d720a2f423ad1003b3ed266b6296a8e559e7