chrome: Make ResetState reset Lacros too
Recently we started providing a lacros.ResetState function and calling
it in a few fixture Reset functions to improve test isolation. This is
working very well. It thus makes sense to call this function in
chrome.ResetState, which is already used by many fixtures and tests, so
that we don't need to add lacros.ResetState calls to countless places.

To be able to use lacros.ResetState in chrome.ResetState, we have to
work around the cyclic import limitation of Go. We do this by
introducing a lacros package internal to chrome containing
lacros.ResetState and anything needed for that. These definitions are
then again reexported in the public packages that currently define them.

BUG=b:245322659,chromium:1314873,b:246497318,chromium:1360293
TEST=../platform/tast/fast-build.sh -T

Change-Id: I14a890f1cc1cadf401988ea3822c8157069b9906