featured: update config installation directory
Currently platform-features.json installs to /etc/init. This is the
improper directory for config files to be.

Update configuration so platform-features.json installs to
the /etc/featured/ directory, a properly dedicated directory for
featured config files.

Additional details for TEST: stop featured, start featured, check /var/log/messages for a dbus "Failed to call method" error due to an unexpected feature name. We expect this error since the feature is not known to chrome.

BUG=b:219758077
TEST=verified that featured starts by checking /var/log/messages on DUT
TEST=tast run ${DUT} security.SELinuxFilesSystem

Cq-Depend: 3905755, 3905759
Change-Id: I4943365ffaf7c91f123b24d6b4e408efb1508692