[TAST] Allow navigating to a WiFi network page.
This change updates the existing method that handles navigating to the
detailed view for a particular Cellular network to support navigating to
the detailed view for a particular WiFi network.

One thing to note is that this change puts the burden of making sure the appropriate technologies are enabled to allow navigating to the detailed view.

BUG=b:245560680
TEST=Manually verified with a temporary test.

Change-Id: I448fa89724cab8030426b72522fe518f1ab31e84