tast-tests: Increase timeout for App Management in arc.LinkCapturing
The test arc.LinkCapturing is commonly failing with a timeout when
interacting with App Management. In local testing, it's common for App
Management to take several seconds to load, so it's likely this is just
taking too long and hitting the timeout, without any particular error.

This CL increases the timeout to 30 seconds, and also adds a specific UI
tree dump if this interaction fails, so that we have more debugging
information if this continues failing.

BUG=chromium:1353271
TEST=tast run $DUT arc.LinkCapturing.vm

Change-Id: Idf71a2c118914ed6593e57df4d3d91387117ed5e