hwsec: Fix intermittent failures in TPMContest
hwsec.TPMContest has been intermittently failing because of the
nondeterministic time it takes to generate an RSA key. Calls to
oemcrypto_hw_ref_tests can timeout waiting to use the TPM.
Also, the test is consistently failing on non-AMD based systems.

Remove the calls to tpm-manager and trunks_client that exercise the
x86<->tpm interface and replace them with a call to read the ph_enable
sysfs descriptor which is speedy, and a better candidate to repeatedly
stress locking mechanisms and/or race conditions that could occur
with the psp.

Although the test should ideally be able to run on all architectures, limit
the test to duts with an AMD cpu for now.

BUG=b:240975458
TEST=tast run <DUT_IP> hwsec.TPMContest

Cq-Depend: chromium:3927898, chromium:3928384
Change-Id: Ib0ac43937c9941f983044782c7ed8c8a5a2c14a6