Projector: Refactor FixtData to use common interfaces
This CL refactors the familylink and projector FixtData structs to
implement common interfaces such as HasChrome and HasFakeDMS. It also
adds two new interfaces HasTestConn and HasPolicyUser. By implementing
interfaces, we can mix and match different types of fixtures in tests
(for example, we can use both familylink and projector fixtures in the
same test).

BUG=b:244236866
TEST=tast run -buildbundle=cros $DUT familylink.* projector.*

Change-Id: Id99318752573ad2f34a61ceb08654da3ab0d1d1a