camera: Remove workaround logic for unexpected minimization in CCAUIIntent
Previously, we hit issue that the test app might be unexpectedly
minimized when running after some arc.WindowState.* tests.

And to workaround it, we add a logic to bring the test app to front when
the UI element is not shown within timeout.

Since now we use our own Fixture, ideally we don't be affected by these
tests. As a result, this CL removes the workaround.

BUG=b:248426090
TEST=tast run {DUT} camera.CCAUIIntent*

Change-Id: I599644a31dda004dfa34f2ebe9e9033b4b0585cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3934695
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Wei Lee <wtlee@chromium.org>
Reviewed-by: Dean Liao <deanliao@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>