tast-tests: Refactor arc.ImagePaste.* to use the web page
This CL make arc.ImagePaste test stop using the custom extension
to show the image.
By this change, the test can use the fixtures now.
Also this CL removes the hack to click the button on the web page
which is needed because uiauto.Finder is not working correctly
for the custom extension window.

BUG=b:247034953
TEST=tast run $DUT arc.ImagePaste.vm arc.ImagePaste.lacros_vm

Change-Id: Ic918332fce666b58e8b4d2f73d46cdb5fd78da20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3929982
Tested-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>