arc: Adjust ArcFileReaderTest to App Service file task handling
Currently, file tasks for ARC apps are found using intent_helper mojo
methods. This implementation, however, is planned to be replaced with
a new one that uses App Service. The new implementation requires that:
1. The app to be registered to the launcher.
2. There is an intent filter just uses mime types.
This CL adjusts ArcFileReaderTest to this new implementation.
Note that we just add the text/plain mime filter, because the issue
reported in b:159522677 does not seem to be observed anymore.

BUG=b:249297999
TEST=Run arc.MyFiles and arc.Drivefs with and without the new
TEST=implementation enabled.

Change-Id: Ief4c4abb19c50ecdf97585fb68f5c76a0c1d65d8