ALSAConformance: Relex rate_error limit for grunt devices
Relex the rate_error limit for grunt devices to fix the test flakiness. We observed that grunt devices need rateCriteria = 0.15 to pass the test steadily.
As the rate diff is minor for grunt, we relax the
criteria first to detect the most serious problems.

BUG=b:136614687, b:234534871
TEST=tast run $DUT audio.ALSAConformance

Change-Id: Ic12a41858dd20b12072015112e53e0e8d08999c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3921096
Reviewed-by: Dean Liao <deanliao@chromium.org>
Tested-by: Judy Hsiao <judyhsiao@google.com>
Commit-Queue: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>