tast-tests: get WiFi service path before connection
With this change test are now able to get the service path before
connecting to it.  In addition you can now also query properties of the
service pointed by that path (instead of just the selected one). This
allows checking how properties of the service changes when new endpoints
are added to it.

BUG=b:226138205
TEST=Make sure the tast client builds correctly (and possibly run
     FgsecMultiConnect test from the next CL).

Change-Id: I8d76262a4ec8bf92eddfc4cdf61d86852e7b15a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3930138
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>