tast-tests: Create utility that takes screenshots at fixed intervals
Create a ScreenshotRecorder that takes screenshots of the device
throughout the execution of cujrecorder.Recorder.Run. The utility also
takes a screenshot at the time Stop is called, to ensure we see the
state of the device at the end of the test. Each screenshot is stored
as a jpg file in a folder called "cuj-screenshots".

This utility is currently only active on ui.TabSwitchCUJ*, to ensure
that it works properly before expanding it to other tests.

BUG=b:240575502
TEST=tast run <DUT> ui.TabSwitchCUJ, and ensure that there is a folder
created called "cuj-screenshots" in the out directory, and that the
screenshots match what I saw during the test.

Change-Id: I29f9350d14f5fd6cc328ffd44c9bf4dedbbcbb51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3929835
Commit-Queue: Navin Ramsaroop <ramsaroop@google.com>
Tested-by: Navin Ramsaroop <ramsaroop@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>