tast-tests: Create utility that takes screenshots at fixed intervals
Create a ScreenshotRecorder that takes screenshots of the of the device
throughout the execution of cujrecorder.Recorder.Run. The utility also
takes a screenshot at the time Stop is called, to ensure we see the
state of the device at the end of the test. Each screenshot is stored
as a jpg file and zipped to lessen their file sizes.

This utility is currently only active on ui.TabSwitchCUJ*, to ensure
that it works properly before expanding it to other tests.

BUG=b:240575502
TEST=tast run <DUT> ui.MeetCUJ*, and ensure that the only new file in
the out directory is test-screenshots.zip.

Change-Id: I29f9350d14f5fd6cc328ffd44c9bf4dedbbcbb51