cellular: Add utility to tag known bugs on variants and boards
It's very common on cellular tests that some tests will always fail on
specific variants or boards. Some will fail for a long time or short
time, but having these known failures on stainless and tastmon makes it
very hard to focus on regressions. Add a tag on known bugs that can
easily be filtered out on stainless.

BUG=None
TEST=tast run tast_lab_hoglin_amarisoft  cellular.ModemFWManifestVerification

Change-Id: If626d1eab270b61627bad5129f0ac16b27bbaa94