typecd: Convert cleanup err to log in fixture
In commit adbfddf3e674 ("typec: Don't fail the test on CCD keepalive
setting failure") we had converted a cleanup error message to log, since
a failure of the servo to bring its servod interfaces back up shouldn't
be attributed to the test itself.

Unfortunately, that change got dropped when we switched to using
fixtures. Re-introduce that change, so that failures which aren't
related to any Type-C functionality aren't miscategorized as
typec.Basic failures.

BUG=b:190578402
TEST=typec.Basic passes 20/20 times on volteer.

Change-Id: I3ca407c14cbb9034cdda7f182a73903530d9e31e