[TAST] Add test to connect to WiFi via UI.
This change adds a test to cover connecting to a WiFi network via
clicking on the network in the Quick Settings. In order to facilitate
adding this test we needed a Quick Settings service and an OS Settings
service, and instead of splitting this into a separate change where they
would be added without being used they were included here.

BUG=b:244330490,b:244776362
TEST=tast run ... *ConnectFromQuickSettings*

Change-Id: I2b81a6142e94df227a79334faa7f09dd2380684d