tast-tests: Deflake personalization.SetDarkLightWallpaper
While daily refresh is enabled, e.g. http://screen/BKVrmZDyYcVr568, the
test will fail because "Daily Refresh" will be appended to the UI node.
To deflake the test, we need to loosen the condition in the API
wallpaper.CurrentWallpaperWithSpecificNameFinder by adding an OR
condition such that the current wallpaper node could be found regardless
if Daily Refresh is enabled.

BUG=b:243452106
TEST=tast run $DUT personalization.SetDarkLightWallpaper

Change-Id: I8de0f0500ae971da925cbaed8ebeb4717b87193b