tast-tests: Add /run/lacros to prunePaths
Lacros's squashfs is mounted in /run/lacros.
It remains mounted after ui restart, so depending on the sequence
in which tests are executed, security.ToolchainOptions will scan it,
searching for ELF files.

Currently, Lacros's nacl binary is being compiled with the
Chrome toolchain, not the hardened one, which is causing the
test to fail. Work is in progress to change this, but for now,
this commit adds the binary to the allowlist.

BUG=chromium:1366161
TEST=Verified test passes even after running Lacros.

Change-Id: Ic3acad2faf7bbd2acd9484f46945009cad234f56