tast-tests: Update shelf.AppRearrangement test
With the shelf-drag-to-pin feature landed, now the unpinned apps can
be dragged across the separator on the shelf and get pinned. This
conflict with the current test flow in AppRearrangement, which supposed
that the unpinned apps can not be pinned by dragging. This cl updates
the test to match the current expected behavior.

Note that the previous flaky issue before the feature is landed is not
fixed in this cl.

BUG=chromium:1370347
TEST=tast unit tests passed. ran shelf.AppRearrangement on a volteer
device.

Change-Id: Id40e7189d5057acc20ccdc75056a6ada308da7f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3933869
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Wen-Chien Wang <wcwang@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Commit-Queue: Wen-Chien Wang <wcwang@chromium.org>