tast-tests: Randomly select from copies of drive documents in CUJ tests
Too many users on a single document leads to Google Docs, Sheets, or
Slides giving a stripped down ui and feature set for users. This makes
CUJ tests that open these documents inconsistent, because a good score
could just be because we were seeing a lighter version of the webpage.
Thus, we expand on the current method of randomly selecting Google Docs
to include Google Sheets and Google Slides.

BUG=b:248539074
TEST=tast run <DUT> ui.TaskSwitchCUJ.vm ui.DesksCUJ ui.GoogleSlidesCUJ
ui.GoogleSheetsCUJ

Change-Id: I25546363859e7ce3de74254c9fad36d00c4488f6