cellular: Rename smoke test and increase timeout
The smoke test had a very low timeout, and in some cases, the modem
could still be registering due to the previous test.

Renaming the test moves it up the test schedule. While not necessary, it
helps schedule the test before tests that can leave the DUT in an
unusable state. SafetyDance and suspend resume are a few tests that can
leave the DUT unusable, causing smoke to fail.

BUG=b:240479899
TEST=tast run $cr cellular.Smoke

Change-Id: I53a640cb940d48c4171023110c3dd437514ba7d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3935253
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>