launcher: Make HideTabletModeLauncher more robust
launcher.HideTabletModeLauncher was hiding tablet mode launcher by
launching chrome app from the launcher UI. This assumed that the
launcher, and the chrome app item view were visible at the time.
The method actually worked even when launcher was initially hidden, but
in a raoundabout way - it would open the launcher, then click
chrome app to close it (which still would not work unless chrome app is
in visible portion of the launcher).

This should fix launcher.RemoveSuggestedSearchResult flakiness when
trying to hide the launcher. The test is requesting the launcher to be
hidden while showing search result UI, in which case app list item views
are not actionable, and expected to be hidden from the automation tree.
Though, it seems like sometimes some items remain present, which
prevents launcher.OpenExpandedView from hiding search UI using search
key.

To make launcher.HideTabletModeLauncher more robust, use
apps.LaunchApp() to open a browser window, and hide the launcher in
tablet mode. This method does not depend on the current UI state, and
should be more robust.

BUG=chromium:1368375
TEST=tast run <dut> launcher.RemoveSuggestedSearchResult*

Change-Id: I04c5eb99c2f4712009ab9d9c13219c246ba75c05