vpn: Stop checking existence of charon after a test
This is the tast test for crrev.com/c/3893879.
This test doesn't have the poll for waiting the previou charon process
disappears before starting a new process. The removed code is no longer
needed after crrev.com/c/3893879 because crrev.com/3893879 includes the
process to check the existence of the previous charon.

BUG=b:241857231
TEST=tast: tast run $DUT network.VPNConnect.l2tp_ipsec*
TEST=../platform/tast/fast_build.sh -T

Change-Id: Ibe31fde2b8842482513ecfda68033037d5bb4218