crostini: refine resize tests
The current resizing logic is to use keyboard to change the disk size on
slider with arrow keys.

There are some drawbacks:
1, It is too slow, especially on devices with large space. It may take
~1 minute to resize once. This could lead to timeout failures as in
http://b/250715234
2, Though the flakiness with a11y tree has been significantly reduced,
it is still inevitable with a11y refresh delays.

This CL changes the test logic to use JS code to change the slider disk
size, which is fast and independent of a11y tree.

BUG=b:250715234, b:239368868
TEST=tast run <DUT> crostini.Resize*

Change-Id: I1782585f04c76a0dc81765fab8d84aab075354e6