ui: Delay cleanup of ScreenRecorder to capture crashes in recordings
When screen recording with cujrecorder.Recorder, an abrupt crash is not
recorded because the recording finishes too quickly. Add a sleep before
cleanup to ensure any last actions performed by the test are recorded
properly.

BUG=b:249571136
TEST=tast run -var=record=true <DUT> ui.DesksCUJ.lacros, and right after
the last page is opened, I navigate to
chrome://inducebrowsercrashforrealz. The resulting recording should
include footage for navigating to the webage and the browser crashing,
which it did not have before.

Change-Id: I6f0c0fb842cc9f5ab7b7203709121239505c165a