[TAST] Add DumpUITree to ChromeUIService.
This change adds an API to the ChromeUIService that dumps the UI tree to
the output directory for the provided testing context.

When regenerating the .pb.go the file for the modified service it also
regenerated the files for the rest of the services.

BUG=None
TEST=Manually tested with a temporary test.

Change-Id: I751955ec0fa5bd532596f50b064f661f3ffd15cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3936436
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Chad Duffin <chadduffin@chromium.org>
Commit-Queue: Chad Duffin <chadduffin@chromium.org>