power: Skip HidWake on octopus
The HidWake test will currently fail on octopus where sending a HID
event from servo will wake the device, but it gets attributed to an
acpi device. While this gets debugged, tastboard should not continue to
run this test on octopus. This CL will have octopus skip HidWake.

BUG=b:251275943
TEST=running test on octopus results in "power.HidWake [ SKIP ]
PlatformId matched with skip-on list"

Change-Id: Ibc1a20a694485c60026bfffa4d8803c069651198