healthd: Remove the hardcoded number of routines in DiagnosticsList
The number of always-supported routines varies from time to time. Remove
the hardcoded number to reduce the maintenance effort.

This commit affects little test coverage since the main purpose of this
test is to ensure the function can be invoked without crashes. The
content of the response is covered by tests of individual routines.

BUG=b:242832956
TEST=tast run ${DUT_IP} health.DiagnosticsList
TEST=tast/fast_build.sh -T

Change-Id: Ia3fe5b2989048ee17139bd08975d46df3020bd75