[Tast] Metadata update to meta package
Adding bug field and ensuring the owning area's group is the first listed contact.
See go/metadata-tweaks-2022 for more info.

You have been added to this review because you are listed as a contact
for a test in this package.  If it doesn't make sense for you to be
contacted about failures/issues in the future, please leave a comment
below.  Similarly, please comment if the bug component listed is not
where bugs for this test should be initially filed.

BUG=None
TEST=N/A

Change-Id: I363bfcd22d3e7bbe616f959aba33d1d22ee689b2