arc: Run DataCollector on release branches.
For main branch DataCollector runs in PFQ and automatically generates
caches for the each version in GoldenEye. However release branches have
different configuration and don't run PFQ. Prior this CL we had to
manually run DataCollector on release branches and pin version. This is
quite unreliable and as result we missed caches in M105 that led to ANR
rate spike. This automates this flow on release branches.

BUG=b:251230500
TEST=tast run ${DUT} arc.DataCollector.*branch_uprev on various devices.

Change-Id: I01c5b98601cb580c221cf0e919558b3ecda57e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3940035
Tested-by: Yury Khmel <khmel@chromium.org>
Commit-Queue: Yury Khmel <khmel@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Reviewed-by: Alan Ding <alanding@chromium.org>