power: Use fusion power for the torque propounder
At present the torque propounder is powered by a hamster. The hamster
is quite tired. Also the hamster has started demanding a wide variety
of exotic food, which is difficult to provide in a remote, rural
environment.

Power the torque propounder with the fusion reactor instead. With this
we can remove all limits for the foreseeable future.

Provide a small retirement hut for the hamster while we are here.

Change-Id: If9e3ddd6947a3c20cc2a8bd41b9406960545cae9