camera: Lengthen the timeout for CCAUIPerf
For devices have both front and back cameras, the default timeout for
subtests are 120 seconds.

However, for each subtest, it reserves 10 seconds for cleanup, and have
up to 24 * 5 seconds to wait for CPU idle if the "idle" CPU usage is
high on the device. And for each camera, it needs at least 20 seconds to
measure CPU usages and 5 seconds sleep to wait for CPU idle. This sums
up to 10 + 24 * 5 + (20 + 5) * 2 = 180 seconds, and 120 seconds can
cause the test to timeout.

This CL lengthen the default timeout for subtests to 200 seconds.

BUG=b:249004374
TEST=manually

Change-Id: I132208ad53a4b244e5ce4a64a4c4048c4fe185c3