faft: Fix bug in BootBatteryCutoff()
In case that ectool flashprotect fails in changing the wp,
check for the wp status using flashrom. If the status matches
with the expected state, continue the test and log the stderr
in a message for debugging purposes. Also, extend a few timeouts
in checking for charger, and waiting for DUT to reconnect.

BUG=b:200305491
TEST=manual on nautilus, nautiluslte, kasumi360.

Change-Id: I08dd8768fe6fb7ee5ff44add62b7441c38ee92e6