Fix full restore tasts.
Use ComboBoxSelect for the tast test FullRestoreAlwaysRestore.

When the chrome browser window is maximum, one of the browser window
button is called "Restore" as well, which has the same name as the full
restore notification button. So set Ancestor for the Restore button, to
set the search scope for the Restore button in the full restore
notification dialog for the tast test FullRestoreOpenSettings.

BUG=chromium:1348439
TEST=run full restore tast tests

Change-Id: I272d9707cb37b9b1b307361839162b8d64159ab5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3945631
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Nancy Wang <nancylingwang@chromium.org>
Auto-Submit: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>