network: check if WireGuard can configure IPv6 address
This patch updated the tast test to check if an IPv6 address can be
configured for wireguard. In crrev.com/c/397310, shill receives an IPv4
address and IPv6 addresses in a string list and creates IPAddress
objects for overlay IPs and allowed IPs. This patch checks if wireguard
can configure the IPv6 addresses with IPv4 addresses correctly (with
peers) by pinging.

BUG=none
TEST=tast run $DUT network.VPNConnect.wireguard*

Change-Id: I1756a82019aed57d3a7f0d75511e412519676182