audio: Modify tests for Audio Player testing
We want to automate the manual test "Audio Player system app audio
playback testing"
https://testtracker.googleplex.com/testplans/testcase/detail/3524960?id=36098

audio.PlaybackAudioControls and audio.CheckingAudioFormats fit our
needs, but need to be modified to match the requirements in the manual
test.

Add additional file formats in audio.CheckingAudioFormats and add tests
for expanding and collapsing playlist in audio.PlaybackAudioControls.
Also fixed the issue of audio.PlaybackAudioControls failing because of
multiple File app windows open and both tests failing because of demanding
internal speaker to be the playing device.

BUG=b:231367554
TEST=tast run <device IP> audio.PlaybackAudioControls
TEST=tast run <device IP> audio.CheckingAudioFormats

Change-Id: I885c014221c561a07422aa7cb29440190c1b3172