security.SandboxedServices: Detect unexpected CAP_SYS_ADMIN processes.
CAP_SYS_ADMIN is a root-equivalent capability. Have the teset detect
processes running with CAP_SYS_ADMIN that are not accounted for in the
baseline.

Add baselines for the three unaccounted-for processes:
-spaced
-dnsproxyd
-cros-disks

Report the username of a running process so that the test can print
a suggested baseline entry for a failing process.

Also, update license headers.

BUG=b:253265398
TEST=Passes on betty-pi-arc.

Change-Id: I870f01789ca6f9d6bfc9d304425a5849bf8d70a1