security.SandboxedServices: Detect unexpected CAP_SYS_ADMIN processes.
CAP_SYS_ADMIN is a root-equivalent capability. Have the test detect
processes running with CAP_SYS_ADMIN that are not accounted for in the
baseline.

Add baselines for the three unaccounted-for processes:
-spaced
-dnsproxyd
-cros-disks

Report the username of a running process so that the test can print
a suggested baseline entry for a failing process.

Also, update license headers.

BUG=b:253265398
TEST=Passes on betty-pi-arc.

Change-Id: I479d4616fcd7d4e27d35e24f1f67d9cbd506df8d