tast-tests: Set regulatory domain to US in OpenWrt router init
Without the regulatory domain being set to US at least once it
seems that OpenWrt routers default to DFS-UNSET, which breaks
tests that use channels that would otherwise work if it was set
to US. Legacy/Gale routers always set this to US already, and it
is the expected environment for our wifi tests. This change ensures
that OpenWrt routers always have the regulatory domain set to US.

A previous version of the OpenWrt router controller did this with
the OpenWrt UCI command, but that had since been removed when
the OpenWrt base hostapd process was disabled in our custom images.
We believe that this was not caught until a new OpenWrt router was
configured as it had never had any regulatory domain set previously
since it was removed prior to tests being run on it. This change
sets the domain with the iw command just like it does with the
Legacy router controller, not UCI, as it is now possible to do so
now that the base hostapd process is disabled.

BUG=b:250673442
TEST=tast unit tests passed; ran the following tests with the new
testbed that previously had issues successfully:
wifi.SimpleConnect.80211a
wifi.SimpleConnect.80211b

Change-Id: I627a3a5033b4b933b7edb2f0abc3b854b31539c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3950892
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>