tast-tests: Refactor wifi tests to always use generic router controller
The wificell test fixture allows for specific use of each router
controller type. This was used in a few tests to specifically select
legacy pcaps/routers. Thus, running these tests with OpenWrt routers
and pcaps failed since they were not legacy ones. This change removes
the functions that easily allow this undesired behavior and refactors
where they were used to use the "Standard" router interface instead,
allowing these tests to be more flexible and allow both types of
routers to be used.

The AX simple connect test does rely on the AX router type so instead
of using Standard for that one, the casting logic was moved into the
test itself instead so that the helper function could still be removed
from the fixture for consistency. It's the only test that uses it (and
likely will continue to be) so there is no need to have it as a fixture
function anyways.

BUG=b:250673442
TEST=tast unit tests passed; successfully ran the updated test on
an OpenWrt testbed:
wifi.RandomMACAddress

Change-Id: I66c5927e8c76c917837628e307e4676c2f1ed840
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3950893
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>